requirements.txt
django==3.2
Pillow
djangorestframework
djoser
djangorestframework_simplejwt
drf-yasg
